Split ChannelInfo from MeterReaderNode
