+++ pymeterreader/core/__init__.py
@@ -1,4 +1,5 @@
+from pymeterreader.core.channel_info import ChannelInfo
 from pymeterreader.core.meter_reader_node import MeterReaderNode
 from pymeterreader.core.meter_reader_task import MeterReaderTask
 
-__all__ = ["MeterReaderNode", "MeterReaderTask"]
+__all__ = ["MeterReaderNode", "MeterReaderTask", "ChannelInfo"]
